+++ example/lib/widgets/week_view_widget.dart
@@ -1,6 +1,5 @@
 import 'package:calendar_view/calendar_view.dart';
 import 'package:example/model/event.dart';
-import 'package:example/widgets/event_provider.dart';
 import 'package:flutter/material.dart';
 
 class WeekViewWidget extends StatelessWidget {
@@ -14,7 +13,6 @@ class WeekViewWidget extends StatelessWidget {
     return WeekView<Event>(
       key: state,
       width: width,
-      controller: DataProvider.of(context).controller,
     );
   }
 }